rename rms_norm_quant
