+++ paddle/fluid/pir/transforms/xpu/rms_norm_quant_xpu_fuse_pass.h
