+++ paddle/phi/infermeta/multiary.h
@@ -1008,21 +1008,21 @@ PADDLE_API void RAdamInferMeta(const MetaTensor& param,
                                MetaTensor* moment2_out,
                                MetaTensor* master_param_outs);
 
-PADDLE_API void RmsNormInferMeta(const MetaTensor& x,
-                                 const MetaTensor& bias,
-                                 const MetaTensor& residual,
-                                 const MetaTensor& norm_weight,
-                                 const MetaTensor& norm_bias,
-                                 const float epsilon,
-                                 const int begin_norm_axis,
-                                 const float quant_scale,
-                                 const int quant_round_type,
-                                 const float quant_max_bound,
-                                 const float quant_min_bound,
-                                 MetaTensor* out,
-                                 MetaTensor* residual_out,
-                                 MetaTensor* inv_var,
-                                 MetaConfig config = MetaConfig());
+PADDLE_API void RmsNormQuantInferMeta(const MetaTensor& x,
+                                      const MetaTensor& bias,
+                                      const MetaTensor& residual,
+                                      const MetaTensor& norm_weight,
+                                      const MetaTensor& norm_bias,
+                                      const float epsilon,
+                                      const int begin_norm_axis,
+                                      const float quant_scale,
+                                      const int quant_round_type,
+                                      const float quant_max_bound,
+                                      const float quant_min_bound,
+                                      MetaTensor* out,
+                                      MetaTensor* residual_out,
+                                      MetaTensor* inv_var,
+                                      MetaConfig config = MetaConfig());
 
 PADDLE_API void RmspropInferMeta(const MetaTensor& param,
                                  const MetaTensor& mean_square,
