+++ paddle/phi/kernels/CMakeLists.txt
@@ -359,7 +359,7 @@ if(NOT
     AND WITH_MKL))
   list(REMOVE_ITEM kernel_cc "fusion/cpu/fused_layer_norm_avx_kernel.cc")
   list(REMOVE_ITEM kernel_cc "fusion/cpu/self_dp_attention_kernel.cc")
-  list(REMOVE_ITEM kernel_cc "fusion/cpu/rms_norm_avx_kernel.cc")
+  list(REMOVE_ITEM kernel_cc "fusion/cpu/rms_norm_quant_avx_kernel.cc")
 endif()
 
 file(
