+++ paddle/phi/kernels/fusion/cpu/rms_norm_quant_avx_kernel.cc
