+++ paddle/phi/kernels/xpu/rms_norm_quant_grad_kernel.cc
@@ -12,28 +12,28 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "paddle/phi/kernels/rms_norm_grad_kernel.h"
 #include "paddle/phi/backends/xpu/enforce_xpu.h"
 #include "paddle/phi/core/kernel_registry.h"
 #include "paddle/phi/kernels/full_kernel.h"
+#include "paddle/phi/kernels/rms_norm_grad_kernel.h"
 
 namespace phi {
 
 template <typename T, typename Context>
-void RmsNormGradKernel(const Context& dev_ctx,
-                       const DenseTensor& x,
-                       const paddle::optional<DenseTensor>& bias,
-                       const paddle::optional<DenseTensor>& residual,
-                       const DenseTensor& norm_weight,
-                       const paddle::optional<DenseTensor>& norm_bias,
-                       const DenseTensor& inv_var,
-                       const DenseTensor& out_grad,
-                       const float epsilon,
-                       const int begin_norm_axis,
-                       const float quant_scale,
-                       DenseTensor* x_grad,
-                       DenseTensor* norm_weight_grad,
-                       DenseTensor* norm_bias_grad) {
+void RmsNormQuantGradKernel(const Context& dev_ctx,
+                            const DenseTensor& x,
+                            const paddle::optional<DenseTensor>& bias,
+                            const paddle::optional<DenseTensor>& residual,
+                            const DenseTensor& norm_weight,
+                            const paddle::optional<DenseTensor>& norm_bias,
+                            const DenseTensor& inv_var,
+                            const DenseTensor& out_grad,
+                            const float epsilon,
+                            const int begin_norm_axis,
+                            const float quant_scale,
+                            DenseTensor* x_grad,
+                            DenseTensor* norm_weight_grad,
+                            DenseTensor* norm_bias_grad) {
   if (x.numel() == 0) {
     dev_ctx.template Alloc<T>(x_grad);
     if (norm_weight_grad) {
@@ -54,11 +54,11 @@ void RmsNormGradKernel(const Context& dev_ctx,
   }
   if (bias || residual) {
     PADDLE_THROW(common::errors::Unimplemented(
-        "bias or residual is not supported in XPU rms_norm_grad yet"));
+        "bias or residual is not supported in XPU rms_norm_quant_grad yet"));
   }
   if (quant_scale > 0.0f) {
     PADDLE_THROW(common::errors::Unimplemented(
-        "quantization is not supported in XPU rms_norm_grad yet"));
+        "quantization is not supported in XPU rms_norm_quant_grad yet"));
   }
   using XPUType = typename XPUTypeTrait<T>::Type;
   xpu::ctx_guard RAII_GUARD(dev_ctx.x_context());
@@ -143,10 +143,10 @@ void RmsNormGradKernel(const Context& dev_ctx,
 }
 }  // namespace phi
 
-PD_REGISTER_KERNEL(rms_norm_grad,
+PD_REGISTER_KERNEL(rms_norm_quant_grad,
                    XPU,
                    ALL_LAYOUT,
-                   phi::RmsNormGradKernel,
+                   phi::RmsNormQuantGradKernel,
                    float,
                    phi::float16,
                    phi::bfloat16) {}
