+++ paddle/phi/ops/yaml/ops.yaml
@@ -4620,7 +4620,7 @@
   args : (Tensor x, Tensor bias, Tensor residual, Tensor norm_weight, Tensor norm_bias, float epsilon, int begin_norm_axis, float quant_scale, int quant_round_type, float quant_max_bound, float quant_min_bound)
   output : Tensor(out), Tensor(residual_out), Tensor(inv_var)
   infer_meta :
-    func : RmsNormInferMeta
+    func : RmsNormQuantInferMeta
   kernel :
     func : rms_norm_quant
     data_type : x
